Fix #1 Created basic directory structure.
Created basic directory structure with asset folders at the root
level of the repository. This can be changed at some point to
have all these folders in an assets folder if deemed neccesary
but since the site itself will probably be simple for now there
really isn't much point to doing that. The directory structure,
if found inadequate should be discussed in Issue #1 however
until the point that that occurs this is directory structure
stands.

The folders are images, javascript, projects, and stylesheets.
all of these are self obvious besides the projects folder.
The projects folder is meant to include pages dedicated to
talking about individual projects worked on as well as provide
an index page that will be a listing and summary of said projects.