Fix #2 Images and index page added
Xenon Apps Logo and simple favicon created by Josh added. The
index page is the splashpage. Once we have actual index page
content, the page will be renamed to something more appropriate
and can be used as a 404 page possibly. As well as a general
coming soon page for whatever content or features we need a place-
holder for.