Updated change log with commit hash
Added commit hash to changelog for the initial changelog commit it's
going to be difficult to keep the hash on the changes to the log
file tracked because whenever you edit it to add information there
is going to be a new hash. So I will not be keeping track of the
changes to the changelog after this initial release. Also. If you're
using vim for an editor during commit messages, I highly suggest
using :set tw=68 to have your commits wrap nicely. The choice of 68
is simply because that is the character width shown in the commit
comments that are auto-generated when you type git commit.