Fix #4 Created placeholder error documents
Created error documents to handle 400-404 and 500 server errors.
Also updated the index page to have the correct copyright metatag.
An .htaccess file has been made so that if we serve this website
from an apache website we'll have the mappings me need already. Each
page is rather similar and is just the logo, a header and a
paragraph with some slightly humorous quotes along with the error
number itself. On certain pages there is a link to a non-existant
contact page that we'll add in eventually.

Since error pages are hardly seen and can be rather amusing rather
than infuriating with the right message to the user. It'd be nice to
have a neat image besides the logo and background mosaic. If one of
the designers get bored I'd love to see what they come up with.