Light and dark themes
Added css to target the site for a light or dark theme. The light theme
uses the subtlepatterns concrete_wall_2 for the background although
this could easily be changed to a custom pattern from a designer.

Removed inline styling, although this was in place so that page loads
would be very fast and content would be delivered quickly, because of
the small size of the site and pages this is not really a neccesary
optimization. If page loads are very slow later on in the sites lifetime
we can revisit inlining the primary styles.