Green Up Project Details
Added text into the "Our Work" section on the projects index page. Also
threw in some reworked text from the Green Up readme to serve as a
description of the project for Green Up.

Updated the styling a little bit for the projects page as well. It
doesn't really look that great in my opinion, but I'm hoping to pass
this type of work off to @ssmullen