Removed Lorem Ipsom and Added index content
Took out placeholder text and added a small bit of actual content to the
index page. The content itself isn't that greant, but I'm not really
sure what to write here to be honest.

Updated styling a little bit so that we can emphasize whatever text we
want just by throwing the emphasize class onto it (hows that for
semantic class naming)