Removed white background and coloring on company
Removed the white background on the company page and made the picture
much smaller and floated it rightward. Also edited the color scheming
out of the style so that when switching from light to dark the text
still shows up correctly and is inherited from the dark/light css
instead.

Granted that the company page's css should be moved out of the head
itself. We still need more content in the page itself besides the rather
silly placeholder texts for each profile. We need images too.