Added content to contact page. Corrected Typos.
Corrected typos in projects/index.html

Added some content to the contact page to flush it out more. It would be
awesome if javascript somehow allowed us to do a mailing form but
honestly I can't imagine that being possible. Unless javascript can call
up system mail process's, which I don't think it can by default. Maybe
there's a way to hack it but if there is I don't know it currently.

I am running the website on an apache server locally, so I could switch
things to php pretty quickly. Which could be really cool because then we
could dynamically choose the theme (light or dark) for the website.
Which I think might be neat.