Alphabetical ordering of developers. Fancy CSS
Re-ordered the list of the developers to be in alphabetical order.
There's no reason for it one way or the other, but it makes sense to
just throw it in some arbitrary ordering because we don't really have a
hierarchy.

Added some css transitioning for hover/active effects to make the web
page feel more like the user is actually interacting with the page. To
be honest, the page feels pretty dead. It's not drawing me in anywhere
or making me excited. There's text, but there needs to be more. There
needs to be images and bold headlines and sleek design if we want this
site to catch anyone's attention.