Added h2 with name to company.html profile segment
Added an h2 to each of the beginning of the Bio's for each developer.

Also updated the dummy profile text a little bit, but not much. If
anyone wants to put a link on their h2 to their personal website like I
did that's completely fine with me.

One thing that I think needs to change is maybe the background on the
list of developers and the styling of the hyperlinks.