Added black background to member list
Added black background to member list to make it look a bit better. Need
to push this styling up into the two theme files so that it's not always
stuck with a black background because that's going to look weird with
the light theme. Similarally for the header styling and coloring of the
links and text,