Made clickable areas clearer in member list
Added a before:content to the member list so that it calls the users
eyes to it a bit more to get them to understand that, yes, they can cick
on the names and something will happen. It might be beneficial to
underline the names themselves.

Also, does anyone else think everything's too boxy?

What I think could really help with the site's look and feel is to get
some images into the empty area on the right side of most pages. The
company page is really the only one that has content on the whole page
and the other pages are really lacking.

Also, I vote that the company page should be renamed developers or
people or something less... impersonal.