Created Xenon.js and Updated Profile Loader.
Removed old profile loading system (switch statements and whatnot)
replaced with more concisle HTML loading system that uses the data
attribute to indicate which profile within the profiles directory should
be loaded.

Once a profile is loaded, it is not loaded again. This setup allows for
only the profiles that are viewed to get loaded onto the page which
increases page load speed and if we were to place Google Analytics onto
the page in an intelligent manner, we could even find out who the most
favorite profile to view is (my bets on Phelan).

By not loading the profile more than once we save on bandwidth as well.
The only thing that is loaded everytime is the image for each profile,
which currently points to Gravatar images for most of us, however once
we have setup profile images then it is easy enough to modify the
Apache config to setup proper cache headers to the browsers can cache
the images. Further saving the internet from request cloggage.