Updated paragraphs for profiles to not use float.
No more float! It's a pain. Moved images around so that profiles with
multiple text doesn't get futzed with. Although I wonder if I could get
away with using an align right on the image and then left on the
paragraph and seeing if the inline takes care of it so I don't need to
use the width percentages.