Added Sean to list of members.
Adding Sean because he's assisted me a lot with creating the
non-parralax version of this site and picking color schemes that worked
well. Also because whenever I need to know how to make a site look
professional he's the guy I go to.

Updated Ethan and Shelby profiles a bit to fill out text some more.